+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 
 ### New features
 
+- Added `\Wizaplace\SDK\Catalog\ProductAttribute::getValues`
+- Deprecated `\Wizaplace\SDK\Catalog\ProductAttribute::getValueIds`
+- Deprecated `\Wizaplace\SDK\Catalog\ProductAttribute::getValue`
+- Added `\Wizaplace\SDK\Discussion\DiscussionService::startDiscussionWithVendor`
+- Added `\Wizaplace\SDK\Discussion\DiscussionService::startDiscussionFromDeclinationId`
 - Added `\Wizaplace\SDK\Pim\Product\ProductService::getProductById` and associated classes
 
 ### Bugfixes